gimme-words-bot/pyproject.toml
[project]
name = "gimme-words-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.7",
]

[dependency-groups]
dev = [
    "ruff>=0.15.7",
]
